# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/organizations": {
      "post": {
        "operationId": "OrganizationsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOrganizationDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "organizations"
        ]
      },
      "get": {
        "operationId": "OrganizationsController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "organizations"
        ]
      }
    },
...
    "/campaigns/{campaignId}": {
      "get": {
        "operationId": "CampaignsController_get",
        "parameters": [
          {
            "name": "campaignId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "campaigns"
        ]
      },
      "patch": {
        "operationId": "CampaignsController_update",
        "parameters": [
          {
            "name": "campaignId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCampaignDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "campaigns"
        ]
      }
    },
...
    "/campaign-categories": {
      "get": {
        "operationId": "CategoriesController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "campaign-categories"
        ]
      }
    },
...
    "/organizations/{organizationId}/subscription": {
      "get": {
        "operationId": "OrgSubscriptionController_get",
        "parameters": [
          {
            "name": "organizationId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "billing"
        ]
      }
    },
    "/organizations/{organizationId}/subscription/checkout": {
      "post": {
        "operationId": "OrgSubscriptionController_checkout",
        "parameters": [
          {
            "name": "organizationId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubscribeDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "billing"
        ]
      }
    },
...
    "/campaigns/{campaignId}/fraud/assess": {
      "post": {
        "operationId": "CampaignFraudController_assess",
        "parameters": [
          {
            "name": "campaignId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssessEngagementDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "fraud"
        ]
      }
    },
    "/campaigns/{campaignId}/fraud-cases": {
      "post": {
        "operationId": "CampaignFraudController_flag",
        "parameters": [
          {
            "name": "campaignId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FlagFraudDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "fraud"
        ]
      },
      "get": {
        "operationId": "CampaignFraudController_list",
        "parameters": [
          {
            "name": "campaignId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "fraud"
        ]
      }
    },
...
    "/disputes": {
      "post": {
        "operationId": "DisputesController_open",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OpenDisputeDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "disputes"
        ]
      },
      "get": {
        "operationId": "DisputesController_listOwn",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "disputes"
        ]
      }
    },
    "/disputes/{disputeId}": {
      "get": {
        "operationId": "DisputesController_get",
        "parameters": [
          {
            "name": "disputeId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "disputes"
        ]
      }
    },
    "/disputes/{disputeId}/messages": {
      "post": {
        "operationId": "DisputesController_addMessage",
        "parameters": [
          {
            "name": "disputeId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DisputeMessageDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "disputes"
        ]
      }
    },
    "/admin/disputes": {
      "get": {
        "operationId": "AdminDisputesController_listOpen",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "disputes"
        ]
      }
    },
    "/admin/disputes/{disputeId}/review": {
      "post": {
        "operationId": "AdminDisputesController_review",
        "parameters": [
          {
            "name": "disputeId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "disputes"
        ]
      }
    },
    "/admin/disputes/{disputeId}/resolve": {
      "post": {
        "operationId": "AdminDisputesController_resolve",
        "parameters": [
          {
            "name": "disputeId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "disputes"
        ]
      }
    },
    "/admin/disputes/{disputeId}/reject": {
      "post": {
        "operationId": "AdminDisputesController_reject",
        "parameters": [
          {
            "name": "disputeId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "disputes"
        ]
      }
    },
    "/admin/users": {
      "get": {
        "operationId": "AdminController_listUsers",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "admin"
        ]
      }
    },
    "/admin/users/{userId}/suspend": {
      "post": {
        "operationId": "AdminController_suspendUser",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "admin"
        ]
      }
    },
    "/admin/users/{userId}/reactivate": {
      "post": {
        "operationId": "AdminController_reactivateUser",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "admin"
        ]
      }
    },
    "/admin/organizations": {
      "get": {
        "operationId": "AdminController_listOrganizations",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "admin"
        ]
      }
    },
    "/admin/campaigns": {
      "get": {
        "operationId": "AdminController_listCampaigns",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "admin"
        ]
      }
    },
    "/admin/campaign-categories": {
      "post": {
        "operationId": "AdminController_createCategory",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCategoryDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "admin"
        ]
      }
    },
    "/ai/campaign-brief": {
      "post": {
        "operationId": "AiController_campaignBrief",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CampaignBriefDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "ai"
        ]
      }
    },
    "/ai/content-ideas": {
      "post": {
        "operationId": "AiController_contentIdeas",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ContentIdeasDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "ai"
        ]
      }
    }
  },
  "info": {
    "title": "Zeyoo API",
    "description": "Zeyoo backend public and internal API.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateOrganizationDto": {
        "type": "object",
        "properties": {}
      },
...
      "UpdateCampaignDto": {
        "type": "object",
        "properties": {}
      },
...
      "SubscribeDto": {
        "type": "object",
        "properties": {}
      },
...
      "AssessEngagementDto": {
        "type": "object",
        "properties": {}
      },
      "FlagFraudDto": {
        "type": "object",
        "properties": {}
      },
      "OpenDisputeDto": {
        "type": "object",
        "properties": {}
      },
      "DisputeMessageDto": {
        "type": "object",
        "properties": {}
      },
      "CreateCategoryDto": {
        "type": "object",
        "properties": {}
      },
      "CampaignBriefDto": {
        "type": "object",
        "properties": {}
      },
      "ContentIdeasDto": {
        "type": "object",
        "properties": {}
      }
    }
  }
}